libhwcomposer: Enable rotator on resolution change
When there is a resolution change while playing YUV enable rotator
In scenarios when 2 YUV layers gets swapped in the zordering
results in interchange of resolution. For TILE formats this can
result in flicker due to hardware limitation.
We need to detect this resolution change for two YUV layers and
engage Rotator to convert TILE to Non TILE format to avoid flicker.

Change-Id: I27d836cb34a18b3f10962821d8589d33d0474c9a